+++ src/renderer/gradient_renderer.rs
@@ -21,8 +21,8 @@ impl Renderer for GradientRenderer {
         let vertical = Vec3::new(0.0, 2.0, 0.0);
         let origin = Vec3::new(0.0, 0.0, 0.0);
         for (x, y, pixel) in imgbuf.enumerate_pixels_mut() {
-            let u = x as f64 / width as f64;
-            let v = (height - y) as f64 / height as f64;
+            let u = x as f32 / width as f32;
+            let v = (height - y) as f32 / height as f32;
             let ray = Ray {
                 origin,
                 direction: corner + horizontal * u + vertical * v,
