+++ src/tracer/hit_record.rs
@@ -1,7 +1,7 @@
 use super::Vec3;
 
 pub struct HitRecord {
-    pub t: f64,
+    pub t: f32,
     pub p: Vec3,
     pub normal: Vec3,
 }
