+++ src/tracer/sphere.rs
@@ -2,11 +2,11 @@ use super::{HitRecord, Hitable, Ray, Vec3};
 
 pub struct Sphere {
     pub center: Vec3,
-    pub radius: f64,
+    pub radius: f32,
 }
 
 impl Hitable for Sphere {
-    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord> {
+    fn hit(&self, ray: &Ray, t_min: f32, t_max: f32) -> Option<HitRecord> {
         let oc = ray.origin - self.center;
         let a = Vec3::dot(ray.direction, ray.direction);
         let b = 2.0 * Vec3::dot(oc, ray.direction);
